# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Kursinfo API",
    "description": "Kursinfo API contains information about course introduction texts and image url for course description",
    "version": "1.0.0"
  },
  "basePath": "/api/kursinfo",
  "produces": ["application/json"],
  "paths": {
    "/_checkAPIkey": {
      "get": {
        "operationId": "checkAPIkey",
        "summary": "Test authentication setup",
        "description": "Secured test endpoint which can be used to validate that a client has a working authentication setup\n\n_(System.checkAPIKey)_",
        "parameters": [],
        "tags": ["system"],
        "responses": {
          "200": {
            "description": "An empty response",
            "schema": {
              "$ref": "#/definitions/UserList"
            }
          },
          "401": {
            "description": "This endpoint is secured - **credentials are missing or invalid**."
          },
          "default": {
            "description": "Unexpected error - internal error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      }
    },
    "/v1/images/status": {
      "get": {
        "operationId": "getUploadedImagesNames",
        "summary": "Get all images names uploaded by user",
        "description": "Get all images names uploaded by user",
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "The requested data",
            "schema": {
              "$ref": "#/definitions/AllUploadedImagesName"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      }
    },
    "/v1/sellingInfo/{courseCode}/": {
      "get": {
        "deprecated": true,
        "operationId": "getSellingTextByCourseCode",
        "summary": "Get selling info for a specific course saved in 2 languages, author and image name",
        "description": "Get selling info for a specific course saved in 2 languages, author and image name",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "course code",
            "required": true,
            "type": "string"
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "The requested data",
            "schema": {
              "$ref": "#/definitions/CourseModel"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      },
      "post": {
        "deprecated": true,
        "operationId": "postSellingTextByCourseCode",
        "summary": "Saves an object in the database",
        "description": "Saves changes to an object in the database",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "courseCode of the sellingText to save",
            "required": true,
            "type": "string"
          },
          {
            "name": "sellingText_en",
            "in": "body",
            "description": "English sellingText to save",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CourseModel"
            }
          },
          {
            "name": "sellingText_sv",
            "in": "body",
            "description": "Swedish sellingText to save",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CourseModel"
            }
          },
          {
            "name": "imageInfo",
            "in": "body",
            "description": "Image name which was saved in a azure blob torage",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CourseModel"
            }
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "A successful save"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["write"]
          }
        ]
      }
    },
    "/v1/courseInfo/": {
      "post": {
        "operationId": "postCourseInfo",
        "summary": "Creates an object in the database",
        "description": "Creates an object in the database",
        "parameters": [
          {
            "name": "courseInfo",
            "in": "body",
            "description": "courseInfo to create",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CourseInfo"
            }
          }
        ],
        "tags": ["v1"],
        "responses": {
          "201": {
            "description": "A successful creation",
            "schema": {
              "$ref": "#/definitions/CourseInfo"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["write"]
          }
        ]
      }
    },
    "/v1/courseInfo/{courseCode}/": {
      "get": {
        "operationId": "getCourseInfoByCourseCode",
        "summary": "Get course info for a specific course",
        "description": "Get course info for a specific course",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "course code",
            "required": true,
            "type": "string"
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "The requested data",
            "schema": {
              "$ref": "#/definitions/CourseInfo"
            }
          },
          "404": {
            "description": "No entry found for that courseCode"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["read"]
          }
        ]
      },
      "patch": {
        "operationId": "patchCourseInfoByCourseCode",
        "summary": "Updates an object in the database",
        "description": "Updates an object in the database",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "course code",
            "required": true,
            "type": "string"
          },
          {
            "name": "courseInfo",
            "in": "body",
            "description": "courseInfo to save",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PartialCourseInfo"
            }
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "A successful update",
            "schema": {
              "$ref": "#/definitions/CourseInfo"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["write"]
          }
        ]
      }
    },
    "/v1/imageInfo/{courseCode}/": {
      "post": {
        "operationId": "postImageInfo",
        "summary": "Saves a random image info in the database",
        "description": "Saves changes to an object in the database",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "courseCode of the imageInfo to save",
            "required": true,
            "type": "string"
          },
          {
            "name": "imageInfo",
            "in": "body",
            "description": "The imageInfo to save",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CourseModel"
            }
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "A successful save"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["write"]
          }
        ]
      }
    },
    "/v1/courseWebLink/{courseCode}/": {
      "post": {
        "deprecated": true,
        "operationId": "postCourseWebLink",
        "summary": "Saves a boolean flag isCourseWebLink in the database",
        "description": "Saves a boolean flag isCourseWebLink in the database: true, false",
        "parameters": [
          {
            "name": "courseCode",
            "in": "path",
            "description": "courseCode of the sellingText to save",
            "required": true,
            "type": "string"
          },
          {
            "name": "isCourseWebLink",
            "in": "body",
            "description": "The isCourseWebLink to save",
            "required": false,
            "schema": {
              "$ref": "#/definitions/CourseModel"
            }
          }
        ],
        "tags": ["v1"],
        "responses": {
          "200": {
            "description": "A successful save"
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        },
        "security": [
          {
            "api_key": ["write"]
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header",
      "scopes": {
        "read": "Read access to data",
        "write": "Write access to data"
      }
    }
  },
  "definitions": {
    "Data": {
      "properties": {
        "id": {
          "type": "string",
          "description": "Unique identifier representing some specific data"
        },
        "name": {
          "type": "string",
          "description": "Description of the data"
        }
      }
    },
    "AllUploadedImagesName": {
      "type": "array",
      "items": "string"
    },
    "CourseModel": {
      "type": "object",
      "required": ["courseCode"],
      "properties": {
        "courseCode": {
          "type": "string",
          "description": "Unique course identifier representing course code, t.e, SF1624"
        },
        "sellingText_en": {
          "type": "string",
          "description": "English course introduction, html text"
        },
        "sellingText_se": {
          "type": "string",
          "description": "Swedish course introduction, html text"
        },
        "imageInfo": {
          "type": "string",
          "description": "The imageInfo to save, t.e., Picture_by_own_choice_SF1624.jpg"
        },
        "sellingTextAuthor": {
          "type": "string",
          "description": "The author's kthId to save"
        }
      }
    },
    "CourseInfo": {
      "type": "object",
      "required": ["courseCode"],
      "allOf": [
        {
          "type": "object",
          "properties": {
            "courseCode": {
              "type": "string",
              "description": "Unique course identifier representing course code, t.e, SF1624"
            }
          }
        },
        {
          "$ref": "#/definitions/PartialCourseInfo"
        }
      ]
    },
    "PartialCourseInfo": {
      "type": "object",
      "properties": {
        "sellingText": {
          "type": "object",
          "properties": {
            "en": {
              "type": "string",
              "description": "Selling text in English."
            },
            "sv": {
              "type": "string",
              "description": "Selling text in Swedish."
            }
          }
        },
        "courseDisposition": {
          "type": "object",
          "properties": {
            "en": {
              "type": "string",
              "description": "The courses disposition in English."
            },
            "sv": {
              "type": "string",
              "description": "The courses disposition in Swedish."
            }
          }
        },
        "supplementaryInfo": {
          "type": "object",
          "properties": {
            "en": {
              "type": "string",
              "description": "Supplementary info regarding the course in English."
            },
            "sv": {
              "type": "string",
              "description": "Supplementary info regarding the course in Swedish."
            }
          }
        },
        "imageInfo": {
          "type": "string",
          "description": "The imageInfo to save, t.e., Picture_by_own_choice_SF1624.jpg"
        },
        "sellingTextAuthor": {
          "type": "string",
          "description": "The author's kthId to save"
        }
      }
    },
    "Error": {
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "fields": {
          "type": "string"
        }
      }
    }
  }
}